rust - update rust examples to match c/python
