+++ Cargo.toml
@@ -3,8 +3,10 @@ members = [
         "rust/libceed",
         "rust/libceed-sys",
         "examples/rust/ex1-volume",
+        "examples/rust/ex1-volume-vector",
         "examples/rust/ex2-surface",
-        "examples/rust/ex3-vector-volume",
-        "examples/rust/ex4-vector-surface",
+        "examples/rust/ex2-surface-vector",
+        "examples/rust/ex3-volume",
+        "examples/rust/ex3-volume-vector",
         "examples/rust/mesh",
 ]
