+++ examples/rust/ex1-volume-vector/src/opt.rs
