+++ examples/rust/ex2-surface-vector/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "ex4-vector-surface"
+name = "ex2-surface-vector"
 version = "0.11.0"
 authors = [
     "Jeremy L Thompson <thompson.jeremy.luke@gmail.com>",
