+++ examples/rust/ex2-surface-vector/src/opt.rs
