+++ examples/rust/ex3-volume-vector/.gitignore
@@ -1,0 +1,2 @@
+target
+Cargo.lock
