+++ examples/rust/ex3-volume-vector/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "ex3-volume-vector"
+version = "0.11.0"
+authors = [
+    "Jeremy L Thompson <thompson.jeremy.luke@gmail.com>",
+]
+edition = "2018"
+
+[dependencies]
+clap = { version = "4.0.17", features = ["derive"] }
+libceed = { path = "../../../rust/libceed" }
+mesh = { path = "../mesh" }
+
+[package.metadata.release]
+release = false
