+++ examples/rust/ex3-volume/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "ex3-vector-volume"
+name = "ex3-volume"
 version = "0.11.0"
 authors = [
     "Jeremy L Thompson <thompson.jeremy.luke@gmail.com>",
